requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.6
django-crispy-forms==1.11.0
django-filter==2.4.0
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
Markdown==3.3.3
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.3
